requirements.txt
graphviz==0.19.2
numpy==1.21.6
pandas==1.3.5
requests==2.27.1
psycopg2==2.9.3
python-dotenv==0.20.0
text-unidecode==1.3
urllib3==1.26.9
websocket-client==1.3.2
pyjwt==2.6.0
python-dotenv==0.20.0